requirements.txt
langchain-community==0.2.1
langchain==0.2.1
langchain-openai==0.1.7
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
scikit-learn==1.5.0
fastapi==0.111.0
requests==2.32.2
pymongo==4.7.2
kubernetes==29.0.0
uvicorn==0.29.0
sentence-transformers==3.0.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
transformers==4.41.2
numpy==1.26.4
psycopg2-binary==2.9.9
psycopg2==2.9.9